Wired the function's path to their respective folders and update button text
- Organized functions into their respective folders for better structure.
- Ensured all functions are now operational.
- Changed button text from "Edit review" to "Save Review" after the user updates their review
