+++ templates/account/reviews/display-reviews.html
@@ -310,7 +310,7 @@
         </div>
     </footer>
 
-    <script src="../../../static/js/reviews.js" type="module"></script>
+    <script src="../../../static/js/modules/reviews.js" type="module"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.12.2/dist/sweetalert2.all.min.js"></script>
